Add top-level Git mailmap
Since 2003, this codebase has substantially grown thanks to the
input of several tens of contributors. Over two decades, commits
from the same author are associated with multiple variants of the
author line. In order to effectively review and audit contributions,
it is necessary to normalize this variant.

This commit attempts to do so by using the Git mailmap functionality
to associate all commits from the same individual to a single
First Name Last Name <email> form. When applicable, either the
University of Dundee or the Glencoe Software email address is used